requirements.txt
python-dateutil>=2.7
numpy>=1.15
matplotlib>=2.0
astropy>=3.0
tqdm>=4.20.0
astroalign>=1.0.3
sep>=1.0.3
PyYAML>=3.13
opencv-python>=3.4.0.12
scikit-image>=0.13.0
